requirements.txt
langchain
modelscope
mdtex2html
streamlit==1.24.0
sentencepiece
accelerate
chromadb
sentence-transformers
unstructured
markdown
pysqlite3-binary
bitsandbytes
gradio==4.4.0